+++ x-twitter-scraper-java-core/src/main/kotlin/com/x_twitter_scraper/api/models/drafts/DraftListResponse.kt
@@ -32,7 +32,9 @@ private constructor(
     private constructor(
         @JsonProperty("drafts") @ExcludeMissing drafts: JsonField<List<Draft>> = JsonMissing.of(),
         @JsonProperty("hasMore") @ExcludeMissing hasMore: JsonField<Boolean> = JsonMissing.of(),
-        @JsonProperty("nextCursor") @ExcludeMissing nextCursor: JsonField<String> = JsonMissing.of(),
+        @JsonProperty("nextCursor")
+        @ExcludeMissing
+        nextCursor: JsonField<String> = JsonMissing.of(),
     ) : this(drafts, hasMore, nextCursor, mutableMapOf())
 
     /**
@@ -231,7 +233,7 @@ private constructor(
      */
     @JvmSynthetic
     internal fun validity(): Int =
-        (drafts.asKnown().getOrNull()?.sumOf { it.validity().toInt() } ?: 0) +
+        (drafts.asKnown().getOrNull()?.sumOf { it.validity() } ?: 0) +
             (if (hasMore.asKnown().isPresent) 1 else 0) +
             (if (nextCursor.asKnown().isPresent) 1 else 0)
