+++ x-twitter-scraper-java-core/src/main/kotlin/com/x_twitter_scraper/api/models/extractions/ExtractionListResponse.kt
@@ -34,7 +34,9 @@ private constructor(
         @ExcludeMissing
         extractions: JsonField<List<ExtractionJob>> = JsonMissing.of(),
         @JsonProperty("hasMore") @ExcludeMissing hasMore: JsonField<Boolean> = JsonMissing.of(),
-        @JsonProperty("nextCursor") @ExcludeMissing nextCursor: JsonField<String> = JsonMissing.of(),
+        @JsonProperty("nextCursor")
+        @ExcludeMissing
+        nextCursor: JsonField<String> = JsonMissing.of(),
     ) : this(extractions, hasMore, nextCursor, mutableMapOf())
 
     /**
@@ -235,7 +237,7 @@ private constructor(
      */
     @JvmSynthetic
     internal fun validity(): Int =
-        (extractions.asKnown().getOrNull()?.sumOf { it.validity().toInt() } ?: 0) +
+        (extractions.asKnown().getOrNull()?.sumOf { it.validity() } ?: 0) +
             (if (hasMore.asKnown().isPresent) 1 else 0) +
             (if (nextCursor.asKnown().isPresent) 1 else 0)
