+++ x-twitter-scraper-java-core/src/main/kotlin/com/x_twitter_scraper/api/models/monitors/MonitorCreateResponse.kt
@@ -309,7 +309,7 @@ private constructor(
     internal fun validity(): Int =
         (if (id.asKnown().isPresent) 1 else 0) +
             (if (createdAt.asKnown().isPresent) 1 else 0) +
-            (eventTypes.asKnown().getOrNull()?.sumOf { it.validity().toInt() } ?: 0) +
+            (eventTypes.asKnown().getOrNull()?.sumOf { it.validity() } ?: 0) +
             (if (username.asKnown().isPresent) 1 else 0) +
             (if (xUserId.asKnown().isPresent) 1 else 0)
