+++ x-twitter-scraper-java-core/src/main/kotlin/com/x_twitter_scraper/api/models/styles/StyleGetPerformanceResponse.kt
@@ -234,7 +234,7 @@ private constructor(
     @JvmSynthetic
     internal fun validity(): Int =
         (if (tweetCount.asKnown().isPresent) 1 else 0) +
-            (tweets.asKnown().getOrNull()?.sumOf { it.validity().toInt() } ?: 0) +
+            (tweets.asKnown().getOrNull()?.sumOf { it.validity() } ?: 0) +
             (if (xUsername.asKnown().isPresent) 1 else 0)
 
     class Tweet
@@ -266,7 +266,9 @@ private constructor(
             @JsonProperty("retweetCount")
             @ExcludeMissing
             retweetCount: JsonField<Long> = JsonMissing.of(),
-            @JsonProperty("viewCount") @ExcludeMissing viewCount: JsonField<Long> = JsonMissing.of(),
+            @JsonProperty("viewCount")
+            @ExcludeMissing
+            viewCount: JsonField<Long> = JsonMissing.of(),
         ) : this(
             id,
             text,
