+++ x-twitter-scraper-java-core/src/main/kotlin/com/x_twitter_scraper/api/models/webhooks/WebhookListResponse.kt
@@ -170,8 +170,7 @@ private constructor(
      * Used for best match union deserialization.
      */
     @JvmSynthetic
-    internal fun validity(): Int =
-        (webhooks.asKnown().getOrNull()?.sumOf { it.validity().toInt() } ?: 0)
+    internal fun validity(): Int = (webhooks.asKnown().getOrNull()?.sumOf { it.validity() } ?: 0)
 
     override fun equals(other: Any?): Boolean {
         if (this === other) {
