+++ x-twitter-scraper-java-core/src/main/kotlin/com/x_twitter_scraper/api/models/x/communities/CommunityRetrieveInfoResponse.kt
@@ -669,7 +669,7 @@ private constructor(
                 (if (moderatorCount.asKnown().isPresent) 1 else 0) +
                 (if (name.asKnown().isPresent) 1 else 0) +
                 (primaryTopic.asKnown().getOrNull()?.validity() ?: 0) +
-                (rules.asKnown().getOrNull()?.sumOf { it.validity().toInt() } ?: 0)
+                (rules.asKnown().getOrNull()?.sumOf { it.validity() } ?: 0)
 
         /** Primary topic */
         class PrimaryTopic
