+++ x-twitter-scraper-java-core/src/main/kotlin/com/x_twitter_scraper/api/models/x/tweets/TweetDetail.kt
@@ -747,7 +747,7 @@ private constructor(
             (if (isNoteTweet.asKnown().isPresent) 1 else 0) +
             (if (isQuoteStatus.asKnown().isPresent) 1 else 0) +
             (if (isReply.asKnown().isPresent) 1 else 0) +
-            (media.asKnown().getOrNull()?.sumOf { it.validity().toInt() } ?: 0) +
+            (media.asKnown().getOrNull()?.sumOf { it.validity() } ?: 0) +
             (quotedTweet.asKnown().getOrNull()?.validity() ?: 0) +
             (if (source.asKnown().isPresent) 1 else 0)
 
@@ -833,8 +833,9 @@ private constructor(
          * Used for best match union deserialization.
          */
         @JvmSynthetic
-        internal fun validity(): Int =
-            additionalProperties.count { (_, value) -> !value.isNull() && !value.isMissing() }
+        internal fun validity(): Int = additionalProperties.count { (_, value) ->
+            !value.isNull() && !value.isMissing()
+        }
 
         override fun equals(other: Any?): Boolean {
             if (this === other) {
@@ -1272,8 +1273,9 @@ private constructor(
          * Used for best match union deserialization.
          */
         @JvmSynthetic
-        internal fun validity(): Int =
-            additionalProperties.count { (_, value) -> !value.isNull() && !value.isMissing() }
+        internal fun validity(): Int = additionalProperties.count { (_, value) ->
+            !value.isNull() && !value.isMissing()
+        }
 
         override fun equals(other: Any?): Boolean {
             if (this === other) {
